Implement the mute and unmute commands, implement moderator permission validation

+++ holobot/extensions/moderation/managers/permission_manager.py
@@ -14,10 +14,26 @@ class PermissionManager(IPermissionManager):
         assert_not_none(server_id, "server_id")
         assert_not_none(user_id, "user_id")
 
+        if permissions == ModeratorPermission.NONE:
+            return
+
         await self.__repository.add_user_permissions(server_id, user_id, permissions)
     
     async def remove_permissions(self, server_id: str, user_id: str, permissions: ModeratorPermission) -> None:
         assert_not_none(server_id, "server_id")
         assert_not_none(user_id, "user_id")
 
+        if permissions == ModeratorPermission.NONE:
+            return
+
         await self.__repository.remove_user_permissions(server_id, user_id, permissions)
+    
+    async def has_permissions(self, server_id: str, user_id: str, permissions: ModeratorPermission) -> bool:
+        assert_not_none(server_id, "server_id")
+        assert_not_none(user_id, "user_id")
+
+        if permissions == ModeratorPermission.NONE:
+            return True
+        
+        owned_permissions = await self.__repository.get_user_permissions(server_id, user_id)
+        return permissions in owned_permissions
